requirements.txt
Flask==3.0.3
python-dotenv==1.0.1
psycopg2-binary==2.9.9
Werkzeug==3.0.3
redis==5.0.4
gunicorn==21.2.0